add(HVRDialogAdapter, HVRDialog): Dialog interaction from Daedalus externals and UI component clicks is now moved to IInteractionAdapter for HVR. XRIT Dialog is broken from now on.
