+++ Assets/Gothic-UnZENity-Core/Scripts/Context/GUZContext.cs
@@ -10,6 +10,7 @@ namespace GUZ.Core.Context
     public static class GuzContext
     {
         public static IInteractionAdapter InteractionAdapter { get; private set; }
+        public static IDialogAdapter DialogAdapter { get; private set; }
 
         public enum Controls
         {
@@ -43,6 +44,7 @@ namespace GUZ.Core.Context
 #if GUZ_HVR_INSTALLED
             Debug.Log("Selecting Context: VR - HurricaneVR");
             InteractionAdapter = new HVRInteractionAdapter();
+            DialogAdapter = new HVRDialogAdapter();
 #else
             Debug.LogError("HVR isn't activated inside Player Settings. Please do before you use it.");
 #endif
