+++ Assets/Gothic-UnZENity-Core/Scripts/Context/IDialogAdapter.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 9b2ffd8ae1a4420899638a70ed01af97
+timeCreated: 1722079897
