+++ Assets/Gothic-UnZENity-HVR/Resources/HVR/Prefabs/VRPlayer-Elements/Menu-Dialog.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3b8af67e66ca2c543a0476bac77da182
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
