+++ Assets/Gothic-UnZENity-HVR/Scripts/HVRDialogAdapter.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 432345d114134762899bd04aa45aa66c
+timeCreated: 1722079706
