+++ Assets/Gothic-UnZENity-XRIT/Scripts/Manager.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 073f98baa00f4867a360177214c91f36
+timeCreated: 1722064203
